feat: add grid overview screens for decks and notes
